Deleted unused AvailableHoursDto. Moved files in TimeEntries to TimeRegistration and deleted TimeEntries folder. Deleted HalfHour as validator. (Replaced with QuarterHour for more consistancy). Deleted SalaryRepository (unused).
