+++ packages/api/AlvTime.Business/TimeRegistration/TimeRegistrationService.cs
@@ -7,7 +7,6 @@ using AlvTime.Business.Interfaces;
 using AlvTime.Business.Options;
 using AlvTime.Business.Overtime;
 using AlvTime.Business.Payouts;
-using AlvTime.Business.TimeRegistration.TimeEntries;
 using AlvTime.Business.Users;
 using AlvTime.Business.Utils;
 using Microsoft.Extensions.Options;
