+++ packages/api/AlvTimeWebApi/Responses/PayoutsResponse.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace AlvTimeWebApi.Responses
 {
