+++ packages/api/Tests/UnitTests/Payouts/PayoutServiceTests.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using AlvTime.Business.Options;
 using AlvTime.Business.Payouts;
 using AlvTime.Business.TimeRegistration;
-using AlvTime.Business.TimeRegistration.TimeEntries;
 using AlvTime.Business.Users;
 using AlvTime.Business.Utils;
 using AlvTime.Persistence.DatabaseModels;
