+++ packages/api/Tests/UnitTests/TimeRegistrationServiceTests.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using AlvTime.Business.Options;
 using AlvTime.Business.Payouts;
 using AlvTime.Business.TimeRegistration;
-using AlvTime.Business.TimeRegistration.TimeEntries;
 using AlvTime.Business.Users;
 using AlvTime.Business.Utils;
 using AlvTime.Persistence.DatabaseModels;
